22.02.17 투수 통산 이닝 계산 변경

+++ SFG/src/main/java/com/SFG/player/model/PitcherStat.java
@@ -17,7 +17,7 @@ public class PitcherStat {
 	private int games;
 	private int game_started;
 	private int saves;
-	private int innings_pitched;
+	private double innings_pitched;
 	private int hits;
 	private int walks;
 	private int strikeouts;
